refactor(ui): break compound component import cycles
Each root file attaches its compounds as static properties, and every
compound imported that root back to reach the context hook, so table,
dropdown, dialog-modal and form were each cyclic. It worked only
because function declarations hoist and the hook is called at render
time; a value read at module scope would have resolved to undefined
under one module evaluation order and not another.

Context and hook now live in a sibling use-<name>.ts that imports
nothing local. DialogModalRef moves with its context, so its consumers
import it from there.
