+++ src/common/presentation/ui/form/compounds/input-wrapper/input-wrapper.tsx
@@ -1,7 +1,7 @@
 import type { ComponentProps } from 'react';
 import { twMerge } from 'tailwind-merge';
 
-import { useForm } from '../../form';
+import { useForm } from '../../use-form';
 
 type FormInputWrapperProps = ComponentProps<'div'>;
