+++ src/common/presentation/ui/form/compounds/select/select.tsx
@@ -10,7 +10,7 @@ import { twMerge } from 'tailwind-merge';
 import type { InputVariants } from '@/ui/variants/input';
 import { inputVariants } from '@/ui/variants/input';
 
-import { useForm } from '../../form';
+import { useForm } from '../../use-form';
 import { InputErrorText } from '../input/error-text/error-text';
 import { InputLabelText } from '../input/label-text/label-text';
