+++ src/common/presentation/ui/table/compounds/values-filter/use-values-filter.ts
@@ -1,6 +1,6 @@
 import type { ChangeEvent } from 'react';
 
-import { useTable } from '@/ui/table/table';
+import { useTable } from '@/ui/table/use-table';
 
 type UseValuesFilterParams = {
   columnId: string;
