+++ src/common/presentation/ui/table/use-column-filter-state.ts
@@ -1,4 +1,4 @@
-import { useTable } from '@/ui/table/table';
+import { useTable } from '@/ui/table/use-table';
 
 export function useColumnFilterState(columnId: string) {
   const { table } = useTable();
