+++ src/module/car/ownership/presentation/ui/modals/promote/promote.tsx
@@ -2,8 +2,8 @@ import type { RefObject } from 'react';
 
 import { Button } from '@/ui/button/button';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 type PromoteModalProps = {
   canPromote: boolean;
