+++ src/module/car/ownership/presentation/ui/sections/ownerships/controls/controls.tsx
@@ -3,7 +3,7 @@ import { useRef } from 'react';
 import { AddButton } from '@/car/ownership/presentation/ui/buttons/add/add';
 import { AddModal } from '@/car/ownership/presentation/ui/modals/add/add';
 import { DashboardSection } from '@/dashboard/ui/section/section';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 type SectionControlsProps = {
   carId: string;
