+++ src/module/car/presentation/ui/cars-gallery/cars-gallery.tsx
@@ -6,7 +6,7 @@ import { AddCard } from '@/car/presentation/ui/cards/add/add';
 import { AddModal } from '@/car/presentation/ui/modals/add/add';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 import { Spinner } from '@/ui/decorative/spinner/spinner';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 import { CarCard } from '../cards/car';
 import { useCarsGallery } from './use-cars-gallery';
