+++ src/module/car/presentation/ui/modals/add/add.tsx
@@ -2,8 +2,8 @@ import type { RefObject } from 'react';
 
 import { AddForm } from '@/car/presentation/ui/forms/add/add';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 type AddModalProps = {
   ref?: RefObject<DialogModalRef | null>;
