+++ src/module/car/presentation/ui/modals/edit/edit.tsx
@@ -3,8 +3,8 @@ import type { RefObject } from 'react';
 import type { CarDto } from '@/car/application/dto/car';
 import { EditForm } from '@/car/presentation/ui/forms/edit/edit';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 type EditModalProps = {
   car?: CarDto;
