+++ src/module/car/service-log/presentation/ui/modals/add/add.tsx
@@ -1,8 +1,8 @@
 import type { RefObject } from 'react';
 
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 import { AddForm } from '../../forms/add/add';
