+++ src/module/car/service-log/presentation/ui/modals/edit/edit.tsx
@@ -2,8 +2,8 @@ import type { RefObject } from 'react';
 
 import type { ServiceLogDto } from '@/car/service-log/application/dto/service-log';
 import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 import { EditForm } from '../../forms/edit/edit';
