+++ src/module/car/service-log/presentation/ui/sections/service-logs/controls/controls.tsx
@@ -3,7 +3,7 @@ import { useRef } from 'react';
 import { AddButton } from '@/car/service-log/presentation/ui/buttons/add/add';
 import { AddModal } from '@/car/service-log/presentation/ui/modals/add/add';
 import { DashboardSection } from '@/dashboard/ui/section/section';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
 
 type SectionControlsProps = {
   carId: string;
