+++ src/module/car/service-log/presentation/ui/tables/service-logs/actions-dropdown/content/use-content.ts
@@ -3,8 +3,8 @@ import { useRef } from 'react';
 
 import { serviceLogRemoveMutationOptions } from '@/car/service-log/presentation/tanstack/mutation/remove';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
-import { useDropdown } from '@/ui/dropdown/dropdown';
+import type { DialogModalRef } from '@/ui/dialog-modal/use-dialog-modal';
+import { useDropdown } from '@/ui/dropdown/use-dropdown';
 
 type UseDropdownContentParams = {
   carId: string;
